Run install after creating files
Not sure about the best method of how to do this, or even if this code works, but something along these lines.
Would be a lot better if after creating a project, the packages were installed. Can't think of an instance I want to create a new app and not install the packages. Plus, this would be more similar to angular-cli.
just an idea...